+++ src/main/java/ru/yandex/practicum/filmorate/annotation/IsContainsSpase.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.annotations;
+package ru.yandex.practicum.filmorate.annotation;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
