+++ src/main/java/ru/yandex/practicum/filmorate/annotation/IsCorrectReleaseDate.java
@@ -1,12 +1,10 @@
-package ru.yandex.practicum.filmorate.annotations;
+package ru.yandex.practicum.filmorate.annotation;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 import javax.validation.constraints.NotNull;
 import java.time.DateTimeException;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 
 public class IsCorrectReleaseDate implements ConstraintValidator<IsCorrectLocalData, LocalDate> {
     int year;
