+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/AlreadyCreatedException.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.filmorate.exceptions;
+
+public class AlreadyCreatedException extends RuntimeException {
+
+    public AlreadyCreatedException(String message) {
+        super(message);
+    }
+}
