--- src/main/java/ru/yandex/practicum/filmorate/exceptions/BeanNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exceptions;
-
-public class BeanNotFoundException extends RuntimeException{
-    public BeanNotFoundException(String message){
-        super(message);
-    }
-}
