+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -2,10 +2,12 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.yandex.practicum.filmorate.annotations.IsCorrectLocalData;
+import ru.yandex.practicum.filmorate.annotation.IsCorrectLocalData;
 
 import javax.validation.constraints.*;
 import java.time.LocalDate;
+import java.util.HashSet;
+import java.util.Set;
 
 @Data
 @AllArgsConstructor
@@ -22,5 +24,7 @@ public class Film {
     private LocalDate releaseDate;
     @Positive(message = "Длительность фильма должна быть положительной.")
     private Long duration;
+
+    private final Set<Integer> usersIdWhoLikes = new HashSet<>();
 }
