+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -1,10 +1,11 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.storage.film;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
@@ -13,11 +14,14 @@ import java.util.List;
 public class InMemoryFilmStorage implements FilmStorage {
 
     private final HashMap<Integer, Film> filmStorage = new HashMap<>();
-    private Integer counter = 1;
 
     @Override
     public List<Film> getAllFilms() {
-        return new ArrayList<>(filmStorage.values());
+        if (filmStorage.isEmpty()){
+            return Collections.emptyList();
+        } else {
+            return new ArrayList<>(filmStorage.values());
+        }
     }
 
     @Override
@@ -33,23 +37,13 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film setFilm(Film film) {
-        if (film.getId() == null) {
-            film.setId(generateId());
-        } else {
-            if (film.getId() > counter) {
-                counter = film.getId();
-            }
-        }
         filmStorage.put(film.getId(), film);
         return film;
 
     }
 
     public boolean isThisFilmContained(Integer filmId) {
-        return filmStorage.containsKey(filmId);
+        return filmStorage.get(filmId) != null;
     }
 
-    private Integer generateId() {
-        return counter++;
-    }
 }
