+++ src/test/java/ru/yandex/practicum/filmorate/FilmControllerTests.java
@@ -3,13 +3,13 @@ package ru.yandex.practicum.filmorate;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.validation.annotation.Validated;
-import ru.yandex.practicum.filmorate.controllers.FilmController;
-import ru.yandex.practicum.filmorate.exceptions.BeanAlreadyCreatedException;
-import ru.yandex.practicum.filmorate.exceptions.BeanNotFoundException;
+import ru.yandex.practicum.filmorate.controller.FilmController;
+import ru.yandex.practicum.filmorate.exceptions.AlreadyCreatedException;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.servise.FilmService;
-import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
-import ru.yandex.practicum.filmorate.storage.InMemoryUserStorage;
+import ru.yandex.practicum.filmorate.service.FilmService;
+import ru.yandex.practicum.filmorate.storage.film.InMemoryFilmStorage;
+import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 
 import javax.validation.*;
 import java.time.LocalDate;
@@ -25,9 +25,9 @@ class FilmControllerTests {
     @Test
     void getFilmsTest() {
         InMemoryFilmStorage inMemoryFilmStorage = new InMemoryFilmStorage();
-        FilmService filmService = new FilmService();
         InMemoryUserStorage inMemoryUserStorage = new InMemoryUserStorage();
-        FilmController fc = new FilmController(inMemoryFilmStorage, filmService, inMemoryUserStorage);
+        FilmService filmService = new FilmService(inMemoryUserStorage,inMemoryFilmStorage);
+        FilmController fc = new FilmController(filmService);
         Film aquaMan = new Film(1, "Аквамен", "Фильм про коронование подводного из Пацанов",
                 LocalDate.of(2018, 12, 13), 133L);
         fc.setFilm(aquaMan);
@@ -37,9 +37,9 @@ class FilmControllerTests {
     @Test
     void setFilmsTest() {
         InMemoryFilmStorage inMemoryFilmStorage = new InMemoryFilmStorage();
-        FilmService filmService = new FilmService();
         InMemoryUserStorage inMemoryUserStorage = new InMemoryUserStorage();
-        FilmController fc = new FilmController(inMemoryFilmStorage, filmService, inMemoryUserStorage);
+        FilmService filmService = new FilmService(inMemoryUserStorage,inMemoryFilmStorage);
+        FilmController fc = new FilmController(filmService);
         //check bad request
         Film notCorrectFilmAquaMan = new Film(1, "Аквамен",
                 "Фильм про коронование подводного из Пацанов",
@@ -67,8 +67,8 @@ class FilmControllerTests {
                 "Дата релиза не должна быть ранее 28 декабря 1895 года.");
         notCorrectFilmAquaMan.setReleaseDate(LocalDate.of(2018, 12, 13));
         fc.setFilm(notCorrectFilmAquaMan);
-        final BeanAlreadyCreatedException methodException = assertThrows(
-                BeanAlreadyCreatedException.class, () -> fc.setFilm(notCorrectFilmAquaMan));
+        final AlreadyCreatedException methodException = assertThrows(
+                AlreadyCreatedException.class, () -> fc.setFilm(notCorrectFilmAquaMan));
         assertEquals(methodException.getMessage(),
                 "Film with id=" + notCorrectFilmAquaMan.getId() + " is already created.");
     }
@@ -76,13 +76,13 @@ class FilmControllerTests {
     @Test
     void putFilmsTest() {
         InMemoryFilmStorage inMemoryFilmStorage = new InMemoryFilmStorage();
-        FilmService filmService = new FilmService();
         InMemoryUserStorage inMemoryUserStorage = new InMemoryUserStorage();
-        FilmController fc = new FilmController(inMemoryFilmStorage, filmService, inMemoryUserStorage);
+        FilmService filmService = new FilmService(inMemoryUserStorage,inMemoryFilmStorage);
+        FilmController fc = new FilmController(filmService);
         Film aquaMan = new Film(1, "Аквамен", "Фильм про коронование подводного из Пацанов",
                 LocalDate.of(2018, 12, 13), 133L);
-        final BeanNotFoundException exception = assertThrows(
-                BeanNotFoundException.class,
+        final NotFoundException exception = assertThrows(
+                NotFoundException.class,
                 () -> fc.putFilm(aquaMan));
         assertEquals(exception.getMessage(), "Фильм с id=" + aquaMan.getId() + " не найден.");
         fc.setFilm(aquaMan);
